+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Check our main [developer changelog](https://developer.paddle.com/?utm_source=dx
 
 - Subscription IDs can be omitted when creating customer portal sessions
 - Customer portal session customer ID will always be returned as string
+- `Client.notifications.replay` now calls correct endpoint
 
 ## 1.2.0 - 2024-12-03